re-introduce the setting of kOutOfTime for the spike rejection

+++ RecoLocalCalo/EcalRecProducers/python/ecalGlobalUncalibRecHit_cfi.py
@@ -24,10 +24,10 @@ ecalGlobalUncalibRecHit = cms.EDProducer("EcalUncalibRecHitProducer",
     outOfTimeThresholdGain12mEB    = cms.double(5),      # times estimated precision
     outOfTimeThresholdGain61pEB    = cms.double(5),      # times estimated precision
     outOfTimeThresholdGain61mEB    = cms.double(5),      # times estimated precision
-    outOfTimeThresholdGain12pEE    = cms.double(10),      # times estimated precision
-    outOfTimeThresholdGain12mEE    = cms.double(10),      # times estimated precision
-    outOfTimeThresholdGain61pEE    = cms.double(10),      # times estimated precision
-    outOfTimeThresholdGain61mEE    = cms.double(10),      # times estimated precision
+    outOfTimeThresholdGain12pEE    = cms.double(1000),   # times estimated precision
+    outOfTimeThresholdGain12mEE    = cms.double(1000),   # times estimated precision
+    outOfTimeThresholdGain61pEE    = cms.double(1000),   # times estimated precision
+    outOfTimeThresholdGain61mEE    = cms.double(1000),   # times estimated precision
     amplitudeThresholdEB    = cms.double(10),
     amplitudeThresholdEE    = cms.double(10),
